backend/audio/requirements.txt
websockets
requests
google-genai